move functions at bookingForm to seperate file
